Update SequentialModel.py
Following a Tutorial to learn the basics about tensorflow deep learning models and am now understanding how they work and what I will need to do to my data in order to apply this to my stock project.

It was initially difficult to grasp what each of the concepts meant, and prior to this I had just been guessing and mixing and matching snipets of code i had researched online.

I now understand the purpose of all the steps in my SequentialModel.py code and will be able to apply this knowledge.