Training The Model
I have been messing around with different ways of training the model. I have tried high batch training which yielded results up to 65% accurate, although there was a high variance between epochs so I am not sure how well that method of training actually works. I trained over 10,000 Generations on that model, and am now using that as a base for my small batch training which is more consistent but slower. I am currently seeing results at around 56-58% accuracy adn plan on leaving the Algorithm to train itself for a while and see how it is doing when I return. I have it shuffling the data more often to reduce BIAS and am taking backups of the models to recover if need be.